Move utility check functions into separate file
