+++ optionsfactory/optionsfactory.py
@@ -2,38 +2,6 @@ from collections.abc import Sequence
 from copy import deepcopy
 
 
-# Some common tests that can be imported for clarity
-def is_positive(x):
-    return x > 0
-
-
-def is_positive_or_None(x):
-    if x is None:
-        return True
-    return is_positive(x)
-
-
-def is_non_negative(x):
-    return x >= 0
-
-
-def is_non_negative_or_None(x):
-    if x is None:
-        return True
-    return is_non_negative(x)
-
-
-NoneType = type(None)
-
-
-def with_default(value, default):
-
-    if value is not None:
-        return value
-
-    return default
-
-
 class WithMeta:
     """Type for passing metadata with options value or expression into *OptionsFactory
